requirements.txt
torch
transformers
datasets
peft
pyyaml
scikit-learn
tqdm